Move GeantGeoParams
