+++ src/celeritas/ext/GeantImporter.cc
@@ -71,9 +71,9 @@
 #include "corecel/sys/ScopedMem.hh"
 #include "corecel/sys/ScopedProfiling.hh"
 #include "corecel/sys/TypeDemangler.hh"
+#include "geocel/GeantGeoParams.hh"
 #include "geocel/GeantGeoUtils.hh"
 #include "geocel/ScopedGeantExceptionHandler.hh"
-#include "geocel/g4/GeantGeoParams.hh"
 #include "geocel/g4/VisitVolumes.hh"
 #include "celeritas/Types.hh"
 #include "celeritas/inp/Grid.hh"
