+++ src/geocel/GeantGeoParams.hh
@@ -2,7 +2,7 @@
 // Copyright Celeritas contributors: see top-level COPYRIGHT file for details
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file geocel/g4/GeantGeoParams.hh
+//! \file geocel/GeantGeoParams.hh
 //---------------------------------------------------------------------------//
 #pragma once
