+++ src/geocel/g4/RaytraceImager.cc
@@ -10,9 +10,9 @@
 #include "geocel/rasterize/RaytraceImager.t.hh"
 
 #include "GeantGeoData.hh"
-#include "GeantGeoParams.hh"
 #include "GeantGeoTrackView.hh"
 #include "GeantGeoTraits.hh"
+#include "../GeantGeoParams.hh"
 
 namespace celeritas
 {
