+++ test/celeritas/ext/detail/NaviTouchableUpdater.test.cc
@@ -15,8 +15,8 @@
 #include "corecel/io/Logger.hh"
 #include "corecel/math/ArrayOperators.hh"
 #include "corecel/math/ArrayUtils.hh"
+#include "geocel/GeantGeoParams.hh"
 #include "geocel/UnitUtils.hh"
-#include "geocel/g4/GeantGeoParams.hh"
 #include "geocel/g4/GeantGeoTestBase.hh"
 #include "celeritas/Units.hh"
