+++ test/geocel/GeantImportVolumeResult.cc
@@ -14,8 +14,8 @@
 #endif
 
 #include "corecel/io/Repr.hh"
+#include "geocel/GeantGeoParams.hh"
 #include "geocel/GeantGeoUtils.hh"
-#include "geocel/g4/GeantGeoParams.hh"
 
 namespace celeritas
 {
