+++ test/geocel/g4/GeantGeoTestBase.hh
@@ -6,9 +6,9 @@
 //---------------------------------------------------------------------------//
 #pragma once
 
+#include "geocel/GeantGeoParams.hh"
 #include "geocel/GenericGeoTestBase.hh"
 #include "geocel/g4/GeantGeoData.hh"
-#include "geocel/g4/GeantGeoParams.hh"
 #include "geocel/g4/GeantGeoTrackView.hh"
 #include "geocel/g4/GeantGeoTraits.hh"
