+++ test/geocel/vg/Vecgeom.test.cc
@@ -20,13 +20,13 @@
 #include "corecel/sys/Environment.hh"
 #include "corecel/sys/Version.hh"
 #include "geocel/GeantGdmlLoader.hh"
+#include "geocel/GeantGeoParams.hh"
 #include "geocel/GeantGeoUtils.hh"
 #include "geocel/GeantImportVolumeResult.hh"
 #include "geocel/GenericGeoParameterizedTest.hh"
 #include "geocel/GeoParamsOutput.hh"
 #include "geocel/GeoTests.hh"
 #include "geocel/UnitUtils.hh"
-#include "geocel/g4/GeantGeoParams.hh"
 #include "geocel/rasterize/SafetyImager.hh"
 #include "geocel/vg/VecgeomData.hh"
 #include "geocel/vg/VecgeomParams.hh"
